host_app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pyserial==3.5
python-multipart==0.0.9
